Move `Advanced_Tracking` test suites to `Analytics_4` namespace.
