+++ tests/phpunit/integration/Modules/Analytics_4/Advanced_TrackingTest.php
@@ -1,17 +1,17 @@
 <?php
 /**
- * Class Google\Site_Kit\Tests\Modules\Analytics\Advanced_TrackingTest
+ * Class Google\Site_Kit\Tests\Modules\Analytics_4\Advanced_TrackingTest
  *
- * @package   Google\Site_Kit\Tests\Modules\Analytics
+ * @package   Google\Site_Kit\Tests\Modules\Analytics_4
  * @copyright 2021 Google LLC
  * @license   https://www.apache.org/licenses/LICENSE-2.0 Apache License 2.0
  * @link      https://sitekit.withgoogle.com
  */
 
-namespace Google\Site_Kit\Tests\Modules\Analytics;
+namespace Google\Site_Kit\Tests\Modules\Analytics_4;
 
 use Google\Site_Kit\Context;
-use Google\Site_Kit\Modules\Analytics\Advanced_Tracking;
+use Google\Site_Kit\Modules\Analytics_4\Advanced_Tracking;
 use Google\Site_Kit\Tests\TestCase;
 use Google\Site_Kit\Tests\Modules\Fixed_Blog_Post_Link_Event_List;
 use Google\Site_Kit\Tests\Modules\Dynamic_Blog_Post_Link_Event_List;
